Update ERC223BasicToken.sol
looks like empty should be declared outside of the `if` block so that we have something to return on line 43.

Also want to take advantage of this opportunity to chat with you, @dexaran, and ask for an example contract that implements `ERC223ReceivingContract` trait. Struggling to wrap my head around what kind of validations would be sensible in `tokenFallback` function. Like, how can I ensure that I only accept a specific token, such as `DEX`, but not any other?

While ERC223 is a great idea I'm skeptical of its production adoption without more documentation or, better yet, example contracts.

Thank you!